Made fragments run in background (#28)
Now all fragments of MainActivity run in background. Also removed map redrawing on receiving new current point and fixed possible fatal error when trying to send request to backend using token which can be discarded moments before. This should not really happen, but it's good to be handled. Disabled Back button because it breaks everything and added empty fragment which appears instead of map fragment if there are no permissions because I still can not fid out how to request them without breaking the app.
Fixes #24